change short to uint8_t

+++ src/product/memory.cpp
@@ -5,9 +5,9 @@
  * sets the queue at the proper starting size and creates the correct amount of conveyors
  * @param amount_conveyors how many conveyors that needs to be added
  */
-memory::memory(short amount_conveyors) {
+memory::memory(uint8_t amount_conveyors) {
   _queue.resize(amount_conveyors);
-  for(int i = 0; i < amount_conveyors; i++) {
+  for (uint8_t i = 0; i < amount_conveyors; i++) {
     add_conveyor(); //initialization
   }
 }
@@ -18,8 +18,8 @@ memory::memory(short amount_conveyors) {
  * @param a_brick the brick to be added
  * @param a_conveyor_number which conveyor it gets put on
  */
-void memory::enqueue(brick a_brick, short a_conveyor_number) {
-  if(_queue.size() >= (size_t) a_conveyor_number) { //to ensure proper access
+void memory::enqueue(brick a_brick, uint8_t a_conveyor_number) {
+  if (_queue.size() >= (size_t) a_conveyor_number) { //to ensure proper access
     _queue[a_conveyor_number]->push_back(a_brick);
   }
 }
@@ -29,10 +29,10 @@ void memory::enqueue(brick a_brick, short a_conveyor_number) {
  * shrinks the queue to use less memory
  * @param a_conveyor_number which conveyor to take on from
  */
-brick memory::dequeue(short a_conveyor_number) {
+brick memory::dequeue(uint8_t a_conveyor_number) {
   if (_queue.size() >= (size_t) a_conveyor_number) {
     brick temp_brick = (*_queue[a_conveyor_number])[0];
-    _queue[a_conveyor_number]->erase( _queue[a_conveyor_number]->begin() ); //deletes the first pos and should resize it
+    _queue[a_conveyor_number]->erase(_queue[a_conveyor_number]->begin()); //deletes the first pos and should resize it
     return temp_brick;
   } else {
     return brick::empty_brick();
@@ -51,7 +51,7 @@ void memory::add_conveyor() {
  * @param a_conveyor_number which conveyor to remove
  * \todo this probably needs testing to see if both delete and erase works this way
  */
-void memory::remove_conveyor(short a_conveyor_number) { //not saving any of the deleted data, dont run on existing conveyors
+void memory::remove_conveyor(uint8_t a_conveyor_number) { //not saving any of the deleted data, dont run on existing conveyors
   if(_queue.size() >= (size_t) a_conveyor_number) {
     delete( _queue[a_conveyor_number] ); //this should delete what the pointers point at
     _queue.erase( _queue.begin() + a_conveyor_number ); //this deletes the pointer and resizes the vector
